+++ app/models/design.rb
@@ -352,7 +352,7 @@ class Design < ActiveRecord::Base
   end
 
   def notify_user!(current_user)
-    UserMailer.import_complete(self, current_user).deliver_later if Rails.env.production?
+    UserMailer.import_complete(self, current_user).deliver_later if ENV['emails_enabled'] == 'true'
   end
 
   def insert_new_design_option!(design_option)