+++ app/models/project_user.rb
@@ -12,12 +12,12 @@ class ProjectUser < ActiveRecord::Base
 
   def generate_invite_token!(new_invite_token = Digest::SHA1.hexdigest(Time.zone.now.usec.to_s))
     update invite_token: new_invite_token if respond_to?('invite_token') && invite_token.blank? && ProjectUser.where(invite_token: new_invite_token).count == 0
-    UserMailer.invite_user_to_project(self).deliver_later if Rails.env.production? && !invite_token.blank?
+    UserMailer.invite_user_to_project(self).deliver_later if ENV['emails_enabled'] == 'true' && !invite_token.blank?
   end
 
   private
 
   def notify_user
-    UserMailer.user_added_to_project(self).deliver_later if Rails.env.production? && invite_token.blank? && user
+    UserMailer.user_added_to_project(self).deliver_later if ENV['emails_enabled'] == 'true' && invite_token.blank? && user
   end
 end