+++ test/mailers/user_mailer_test.rb
@@ -1,8 +1,9 @@
 require 'test_helper'
 
+# Tests that mail views are rendered corretly, sent to correct user, and have a
+# correct subject line
 class UserMailerTest < ActionMailer::TestCase
-
-  test "user invited to site email" do
+  test 'user invited to site email' do
     site_user = site_users(:invited)
 
     email = UserMailer.invite_user_to_site(site_user).deliver_now
@@ -13,7 +14,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/#{site_user.creator.name} invited you to Site #{site_user.site.name}/, email.encoded)
   end
 
-  test "user added to project email" do
+  test 'user added to project email' do
     project_user = project_users(:accepted_viewer_invite)
 
     email = UserMailer.user_added_to_project(project_user).deliver_now
@@ -24,7 +25,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/#{project_user.creator.name} added you to Project #{project_user.project.name}/, email.encoded)
   end
 
-  test "user invited to project email" do
+  test 'user invited to project email' do
     project_user = project_users(:pending_editor_invite)
 
     email = UserMailer.invite_user_to_project(project_user).deliver_now
@@ -35,7 +36,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/#{project_user.creator.name} invited you to Project #{project_user.project.name}/, email.encoded)
   end
 
-  test "survey completed email" do
+  test 'survey completed email' do
     sheet = sheets(:external)
 
     email = UserMailer.survey_completed(sheet).deliver_now
@@ -46,7 +47,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/#{sheet.subject.subject_code} completed a survey that you requested for #{sheet.name}\. You can view the completed sheet here:/, email.encoded)
   end
 
-  test "survey user link" do
+  test 'survey user link' do
     sheet = sheets(:external_with_email)
 
     email = UserMailer.survey_user_link(sheet).deliver_now
@@ -57,7 +58,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/Thank you for submitting #{sheet.name}\.\r\n\r\nYou can make changes to your survey responses here:/, email.encoded)
   end
 
-  test "export ready email" do
+  test 'export ready email' do
     export = exports(:one)
 
     email = UserMailer.export_ready(export).deliver_now
@@ -68,7 +69,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/The data export you requested for #{export.project.name} is now ready for download\./, email.encoded)
   end
 
-  test "import complete email" do
+  test 'import complete email' do
     design = designs(:one)
     valid = users(:valid)
 
@@ -80,7 +81,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/The design data import for #{design.project.name} is now complete\./, email.encoded)
   end
 
-  test "daily digest email" do
+  test 'daily digest email' do
     valid = users(:valid)
 
     email = UserMailer.daily_digest(valid).deliver_now
@@ -91,7 +92,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/Dear #{valid.first_name},/, email.encoded)
   end
 
-  test "comment by mail email" do
+  test 'comment by mail email' do
     comment = comments(:one)
     valid = users(:valid)
 
@@ -100,10 +101,10 @@ class UserMailerTest < ActionMailer::TestCase
 
     assert_equal [valid.email], email.to
     assert_equal "#{comment.user.name} Commented on #{comment.sheet.name} on #{comment.sheet.project.name}", email.subject
-    assert_match(/#{comment.user.name} COMMENTED on #{comment.sheet.name} on #{comment.sheet.project.name} located at #{ENV['website_url']}\/projects\/#{comment.sheet.project.id}\/sheets\/#{comment.sheet.id}\./, email.encoded)
+    assert_match(%r{#{comment.user.name} COMMENTED on #{comment.sheet.name} on #{comment.sheet.project.name} located at #{ENV['website_url']}/projects/#{comment.sheet.project.id}/sheets/#{comment.sheet.id}\.}, email.encoded)
   end
 
-  test "project news post email" do
+  test 'project news post email' do
     post = posts(:one)
     valid = users(:valid)
 
@@ -115,7 +116,7 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/This post was added by #{post.user.name} to #{post.project.name} on #{ENV['website_name']}/, email.encoded)
   end
 
-  test "subject randomization" do
+  test 'subject randomization' do
     randomization = randomizations(:one)
     user = users(:valid)
 
@@ -127,4 +128,15 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/#{randomization.subject.name} was randomized to #{randomization.treatment_arm.name} on #{randomization.project.name} by #{randomization.user.name}\./, email.encoded)
   end
 
+  test 'adverse event reported email' do
+    adverse_event = adverse_events(:one)
+    valid = users(:valid)
+
+    email = UserMailer.adverse_event_reported(adverse_event, valid).deliver_now
+    assert !ActionMailer::Base.deliveries.empty?
+
+    assert_equal [valid.email], email.to
+    assert_equal "#{adverse_event.user.name} Reported a Non-Serious Adverse Event on #{adverse_event.project.name}", email.subject
+    assert_match(%r{#{adverse_event.user.name} reported a non-serious adverse event on #{adverse_event.project.name} located here: #{ENV['website_url']}/projects/#{adverse_event.project.to_param}}, email.encoded)
+  end
 end